IoTD/requirements.txt
pandas==0.25.3
paho-mqtt
pycrypto
